+++ automation/src/epilot/models/components/activitytrigger.py
@@ -11,8 +11,6 @@ class Types(str, Enum):
     CREATE_METER_READING = 'CreateMeterReading'
     UPDATE_METER_READING = 'UpdateMeterReading'
     DOC_DOWNLOADED_FROM_PORTAL = 'DocDownloadedFromPortal'
-    MESSAGE_ACTIVITY = 'MessageActivity'
-    SYNC_ACTIVITY = 'SyncActivity'
 
 
 @dataclass_json(undefined=Undefined.EXCLUDE)