+++ automation/src/epilot/models/components/executionstatus.py
@@ -9,3 +9,4 @@ class ExecutionStatus(str, Enum):
     SUCCESS = 'success'
     FAILED = 'failed'
     CANCELLED = 'cancelled'
+    SKIPPED = 'skipped'