+++ automation/src/epilot/models/components/orcondition.py
@@ -8,6 +8,7 @@ from .existscondition import ExistsCondition
 from .numericcondition import NumericCondition
 from .prefixcondition import PrefixCondition
 from .suffixcondition import SuffixCondition
+from .wildcardcondition import WildcardCondition
 from dataclasses_json import Undefined, dataclass_json
 from epilot import utils
 from typing import Dict, List, Optional, Union
@@ -16,6 +17,6 @@ from typing import Dict, List, Optional, Union
 @dataclass_json(undefined=Undefined.EXCLUDE)
 @dataclasses.dataclass
 class OrCondition:
-    dollar_or: Optional[List[Union[OrCondition, Dict[str, List[Union[str, EqualsIgnoreCaseCondition, AnythingButCondition, NumericCondition, ExistsCondition, PrefixCondition, SuffixCondition]]]]]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('$or'), 'exclude': lambda f: f is None }})
+    dollar_or: Optional[List[Union[OrCondition, Dict[str, List[Union[str, EqualsIgnoreCaseCondition, AnythingButCondition, NumericCondition, ExistsCondition, PrefixCondition, SuffixCondition, WildcardCondition]]]]]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('$or'), 'exclude': lambda f: f is None }})